Some more structural changes

+++ src/dimacs.h
@@ -4,5 +4,5 @@
 
 #include "minisat/Solver.h"
 
-int literal_to_dimacs(int lit);
+int literal_to_dimacs(uint64_t lit);
 void to_dimacs(const cnf& c, Solver& S);
